:recycle: refactor test container for DHCP tests
