+++ Makefile
@@ -45,8 +45,8 @@ run: start-db
 
 test: run build-dev
 	./scripts/wait_for_dhcp_server.sh
-	$(DOCKER_COMPOSE) run --rm dhcp-test rspec ./metrics/spec
-	$(DOCKER_COMPOSE) run --rm dhcp-test bash ./dhcp_test.sh
+	$(DOCKER_COMPOSE) run --rm dhcp-test rspec ./spec
+	$(DOCKER_COMPOSE) run --rm dhcp-test bash /dhcp_test.sh
 
 shell: start-db
 	$(DOCKER_COMPOSE) run --rm dhcp-primary sh
