+++ dhcp-service/spec/fixtures/kea_api_config_get_response.json
