+++ dhcp-service/spec/fixtures/kea_api_stat_lease4_response.json
