+++ dhcp-service/spec/fixtures/kea_api_stats_response.json
